// frontend/src/shared/types/toolbar.ts
export type ToolbarButton =
  | 'download'
  | 'save'
  | 'delete'
  | 'crop'
  | 'maximaze'
  | 'turn'
  | 'clip'
  | 'remove'
  | 'user'
  | 'edit'
  | 'reset'
  | 'bold'
  | 'italic'
  | 'fontSize'
  | 'color'
  | 'left'
  | 'center'
  | 'right'
  | 'justify'
  | 'shopping'
  | 'addShopping'
  | 'plus'
  | 'arrowsOut'
  | 'arrowsIn'